+++ payments/WingspanPayments/BenefitsService.cs
@@ -51,10 +51,10 @@ namespace WingspanPayments
     {
         public SDKConfig SDKConfiguration { get; private set; }
         private const string _language = "csharp";
-        private const string _sdkVersion = "4.2.1";
-        private const string _sdkGenVersion = "2.279.1";
+        private const string _sdkVersion = "4.2.2";
+        private const string _sdkGenVersion = "2.287.0";
         private const string _openapiDocVersion = "1.0.0";
-        private const string _userAgent = "speakeasy-sdk/csharp 4.2.1 2.279.1 1.0.0 WingspanPayments";
+        private const string _userAgent = "speakeasy-sdk/csharp 4.2.2 2.287.0 1.0.0 WingspanPayments";
         private string _serverUrl = "";
         private ISpeakeasyHttpClient _defaultClient;
         private Func<Security>? _securitySource;