+++ payments/WingspanPayments/Utils/Utilities.cs
@@ -40,6 +40,7 @@ namespace WingspanPayments.Utils
             {
                 new IsoDateTimeSerializer(),
                 new EnumSerializer(),
+                new FlexibleObjectDeserializer()
             };
         }